pyproject.toml
[project]
name = "Dinogame"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.0",
    "pygame-ce>=2.5.6",
]
